backend/pyproject.toml
[project]
name = "tailormade-backend"
version = "0.1.0"
description = "TailorMade Coloring Book — AI-powered backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "firebase-admin>=6.5.0",
    "fal-client>=0.5.0",
    "anthropic>=0.25.0",
    "boto3>=1.34.0",
    "weasyprint>=62.0",
    "Pillow>=10.3.0",
    "tenacity>=8.3.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "bleach>=6.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
